Multiple bug fixes.
Updates the .nimble file to support a breaking change introduced in 
Nimble 0.13.0.

Locks to a specific blst commit.

Performs a group check against parsed public keys.

Performs a scalar modulus before creating public keys; a requirement 
added a week after this version of the library was initially published 
(and subsequently not noted for ~6 months).